refactor(vestad): merge self_update and update_check into one update module
