+++ vestad/src/update.rs
@@ -1,3 +1,9 @@
+//! Everything about updating the gateway itself: the periodic release check that feeds the
+//! UpdatePill, the phase model and durable ledger that make an update observable, the orchestrator
+//! that drives one update from pre-update snapshots to the restart, and the binary swap it applies.
+
+use std::fmt;
+
 // Poll GitHub for new releases every 5 hours: often enough that the UpdatePill and a
 // power user's manual update see a release promptly. This only governs detection; an
 // auto-update is applied later, by the maintenance cycle in the fleet's 4-5am quiet
@@ -229,6 +235,126 @@ fn snippet(s: &str) -> String {
         Some((end, _)) => format!("{}…", &s[..end]),
         None => s.to_string(),
     }
+}
+
+// --- Binary swap ---
+
+#[derive(Debug)]
+pub enum UpdateError {
+    UnsupportedArch(String),
+    Download(String),
+    Extract(String),
+    Replace(String),
+}
+
+impl fmt::Display for UpdateError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::UnsupportedArch(arch) => write!(f, "unsupported architecture: {arch}"),
+            Self::Download(msg) => write!(f, "failed to download update: {msg}"),
+            Self::Extract(msg) => write!(f, "failed to extract update: {msg}"),
+            Self::Replace(msg) => write!(f, "failed to replace binary: {msg}"),
+        }
+    }
+}
+
+pub struct UpdateOutcome {
+    pub updated: bool,
+    pub restarted: bool,
+    pub current: String,
+    pub latest: String,
+}
+
+/// Downloads the latest vestad binary from GitHub, replaces the current binary,
+/// and restarts the systemd service. Agent code and container restarts are
+/// handled on the next vestad startup.
+/// No-op when the running version is already at or above the latest release on the
+/// given channel. Channel switching never downgrades: a stable channel whose latest
+/// is older than the running (beta) version is simply a no-op.
+pub fn perform_update(channel: crate::channel::Channel) -> Result<UpdateOutcome, UpdateError> {
+    let current = env!("CARGO_PKG_VERSION").to_string();
+    let latest = fetch_latest_tag(channel)
+        .ok_or_else(|| UpdateError::Download("cannot determine latest version".into()))?;
+
+    if !version_less_than(&current, &latest) {
+        tracing::info!(current = %current, latest = %latest, "already up to date, skipping");
+        return Ok(UpdateOutcome {
+            updated: false,
+            restarted: false,
+            current,
+            latest,
+        });
+    }
+
+    tracing::info!(tag = %latest, "starting update");
+
+    update_binary(&latest)?;
+
+    // Don't reinstall the systemd service here — self-replace puts the new
+    // binary at the same path, so the ExecStart line doesn't change. And
+    // reading /proc/self/exe after self-replace returns the path with
+    // " (deleted)" appended, which would break the service file.
+    // The new binary calls ensure_service_installed() on startup if needed.
+
+    let restarted = if crate::systemd::is_active() {
+        tracing::info!("restarting vestad...");
+        if let Err(e) = crate::systemd::restart() {
+            tracing::error!("failed to restart: {e}");
+        }
+        true
+    } else {
+        tracing::info!("updated. run 'vestad' to start.");
+        false
+    };
+
+    Ok(UpdateOutcome {
+        updated: true,
+        restarted,
+        current,
+        latest,
+    })
+}
+
+fn update_binary(tag: &str) -> Result<(), UpdateError> {
+    let target = match std::env::consts::ARCH {
+        "x86_64" => "x86_64-unknown-linux-gnu",
+        "aarch64" => "aarch64-unknown-linux-gnu",
+        other => return Err(UpdateError::UnsupportedArch(other.to_string())),
+    };
+
+    let archive = format!("vestad-{target}.tar.gz");
+    // Download from the resolved tag's release, not the `/latest/` alias: beta tags
+    // are prereleases that `/latest/` never points at, and a pinned tag also avoids
+    // racing a concurrent release that moves `latest`.
+    let url = format!(
+        "https://github.com/elyxlz/vesta/releases/download/v{tag}/{archive}"
+    );
+    let tmp = format!("/tmp/vestad-update-{}", std::process::id());
+    std::fs::create_dir_all(&tmp).ok();
+
+    tracing::info!(tag = %tag, "downloading vestad binary");
+    let status = std::process::Command::new("curl")
+        .args(["-fsSL", "-o", &format!("{tmp}/{archive}"), &url])
+        .status();
+    if !status.is_ok_and(|s| s.success()) {
+        std::fs::remove_dir_all(&tmp).ok();
+        return Err(UpdateError::Download("curl failed".into()));
+    }
+
+    let status = std::process::Command::new("tar")
+        .args(["-xzf", &format!("{tmp}/{archive}"), "-C", &tmp])
+        .status();
+    if !status.is_ok_and(|s| s.success()) {
+        std::fs::remove_dir_all(&tmp).ok();
+        return Err(UpdateError::Extract("tar failed".into()));
+    }
+
+    let new_binary = format!("{tmp}/vestad");
+    self_replace::self_replace(&new_binary).map_err(|e| UpdateError::Replace(e.to_string()))?;
+
+    std::fs::remove_dir_all(&tmp).ok();
+    tracing::info!("vestad binary replaced successfully");
+    Ok(())
 }
 
 #[cfg(test)]
